moves interface into the right namespace.
