+++ src/Entity/CourseObjectiveInterface.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Entity;
 
 use App\Traits\IndexableCoursesEntityInterface;
+use App\Traits\ObjectiveRelationshipInterface;
 
 /**
  * Interface CourseObjectiveInterface
