+++ src/Entity/ProgramYearObjectiveInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
-use App\Traits\IndexableCoursesEntityInterface;
+use App\Traits\ObjectiveRelationshipInterface;
 
 /**
  * Interface ProgramYearObjectiveInterface
